housekeeping: refactor body index screen into separate files
