+++ lib/config/route/app_router.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:gn_accessor/presentation/screens/tools_screen.dart';
 
 import '../../presentation/screens/body_index_form_screen/body_index_form_screen.dart';
-import '../../presentation/screens/body_index_screen.dart';
+import '../../presentation/screens/body_index_screen/body_index_screen.dart';
 import '../../presentation/screens/login_screen.dart';
 import '../../presentation/screens/home_screen.dart';
 import '../../presentation/screens/task_board_screen.dart';
